pyproject.toml
[project]
name = "broad_bros"
version = "0.1.0"
description = ""
authors = [
    {name = "Faculty Science Limited", email = "your.email@faculty.ai"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.4",
    "pydantic-ai>=0.0.14",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "pydantic-logfire>=0.0.1",
]

[[tool.uv.index]]
name = "faculty-artifactory"
url = "https://faculty.jfrog.io/artifactory/api/pypi/faculty-no-pypi/simple"
explicit = true

[tool.uv.sources]
faculty-plot = { index = "faculty-artifactory" }

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "nbstripout>=0.8.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
    "sphinx>=8.1.3",
    "faculty-sphinx-theme>=0.3.1",
  # "faculty-plot==2022.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/broad_bros", "neoops"]

[tool.hatch.version]
source = "vcs"

[project.scripts]
post-install = "pre_commit:install_hooks"

[tool.ruff]
line-length = 120
force-exclude = true
target-version = "py312"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = [
  "E", "F", "W",
  # Include flake8-bugbear
  "B",
  # Include Isort rules
  "I",
  # Prefer absolute imports
  "TID252",
  # Ban print and pprint statements - use a logger
  "T201", "T203",
  # Use loggers correctly
  "LOG001", "LOG002", "LOG007", "LOG009",
  "G001", "G002", "G003", "G004", "G010", "G101",
  "G201", "G202"
]
unfixable = ["B"]

[tool.ruff.lint.isort]
known-first-party = ["broad_bros", "neoops"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
